+++ src/typography/display1/index.tsx
@@ -1,16 +1,16 @@
 import styled from 'styled-components'
 import { color, font, fontWeight } from '_utils/branding'
 
-import TextDisplay2 from './TextDisplay2'
+import TextDisplay1 from './TextDisplay1'
 
-const StyledTextDisplay2 = styled(TextDisplay2)`
+const StyledTextDisplay1 = styled(TextDisplay1)`
   color: ${color.primaryText};
-  font-size: ${font.xxl.size};
-  line-height: ${font.xxl.lineHeight};
+  font-size: ${font.xl.size};
+  line-height: ${font.xl.lineHeight};
   font-weight: ${fontWeight.medium};
 
   &.kirk-text--inverse {
     color: ${color.textWithBackground};
   }
 `
-export default StyledTextDisplay2
+export default StyledTextDisplay1