+++ src/typography/display1/story.tsx
@@ -5,22 +5,22 @@ import { text, boolean, withKnobs } from '@storybook/addon-knobs'
 import { color } from '_utils/branding'
 
 import Section from 'layout/section/baseSection'
-import TextDisplay2 from './index'
+import TextDisplay1 from './index'
 
 const stories = storiesOf('Brand|typography', module)
 stories.addDecorator(withKnobs)
 
-stories.add('TextDisplay2', () => (
+stories.add('TextDisplay1', () => (
   <Section>
     <h1>
-      <TextDisplay2 isInverted={boolean('isInverted', false)}>
+      <TextDisplay1 isInverted={boolean('isInverted', false)}>
         {text('Text1', 'The quick\n brown fox jumps\n over the lazy\n dog')}
-      </TextDisplay2>
+      </TextDisplay1>
     </h1>
     <h1 style={{ backgroundColor: color.primary }}>
-      <TextDisplay2 isInverted>
+      <TextDisplay1 isInverted>
         {text('Text2', 'The quick brown fox jumps over the lazy dog')}
-      </TextDisplay2>
+      </TextDisplay1>
     </h1>
   </Section>
 ))