+++ stories/index.tsx
@@ -1,8 +1,15 @@
 import './stories.css'
 
-// Molecules
+// BRAND
 import '../src/typography/story'
 import '../src/typography/display2/story'
+import '../src/typography/display1/story'
+import '../src/typography/subHeader/story'
+import '../src/typography/subHeaderStrong/story'
+import '../src/typography/title/story'
+import '../src/typography/titleStrong/story'
+import '../src/typography/body/story'
+import '../src/typography/bodyStrong/story'
 
 // Sections
 import '../src/layout/section/baseSection/story'
@@ -12,6 +19,7 @@ import '../src/layout/section/highlightSection/story'
 import '../src/layout/section/tripsSection/story'
 import '../src/layout/section/mediaContentSection/story'
 import '../src/layout/section/columnedContentSection/story'
+import '../src/layout/section/illustratedSection/story'
 
 // Pages
 import '../src/pages/messaging/inbox.story'